Fixed device tree patching on old Armbian releases

+++ dts/armbian/patch_dts.sh
@@ -6,19 +6,21 @@ case "$OVERLAY_MODE" in
   patch)
     TMP_DIR=`mktemp -d`
     for FILE in $FDT_FILE; do
-      dtc -I dtb -O dts -q -o $TMP_DIR/devicetree.dts /boot/dtb/$FILE
+      if [ -e $FILE ]; then
+        dtc -I dtb -O dts -q -o $TMP_DIR/devicetree.dts /boot/dtb/$FILE
 
-      if [ `grep -c -e 'compatible = "pivccu,' $TMP_DIR/devicetree.dts` -eq 0 ]; then
-        echo "piVCCU: Patching DTB $FILE"
-        cp /boot/dtb/$FILE /boot/dtb/$FILE.bak
-        cat $INCLUDE_FILE >> $TMP_DIR/devicetree.dts
-        dtc -I dts -O dtb -q -o $TMP_DIR/devicetree.dtb $TMP_DIR/devicetree.dts
-        cp $TMP_DIR/devicetree.dtb /boot/dtb/$FILE
-      else
-        echo "piVCCU: DTB $FILE was already patched"
+        if [ `grep -c -e 'compatible = "pivccu,' $TMP_DIR/devicetree.dts` -eq 0 ]; then
+          echo "piVCCU: Patching DTB $FILE"
+          cp /boot/dtb/$FILE /boot/dtb/$FILE.bak
+          cat $INCLUDE_FILE >> $TMP_DIR/devicetree.dts
+          dtc -I dts -O dtb -q -o $TMP_DIR/devicetree.dtb $TMP_DIR/devicetree.dts
+          cp $TMP_DIR/devicetree.dtb /boot/dtb/$FILE
+        else
+          echo "piVCCU: DTB $FILE was already patched"
+        fi
+
+        rm $TMP_DIR/devicetree.dts
       fi
-
-      rm $TMP_DIR/devicetree.dts
     done
 
     rm -rf $TMP_DIR
